requirements.txt
Flask>=3.0
Pillow>=10.0
opencv-python
numpy
requests
